+++ R/projection.R
@@ -156,19 +156,23 @@ fit.submodel <- function(x, sigma2, mu, chosen, xt, yt, logistic) {
 #' @param obj Object of class `hsstan`.
 #' @param max.iters Maximum number of iterations (number of predictors selected)
 #'        after which the selection procedure should stop.
+#' @param start.from Vector of variable names to be used in the starting
+#'        submodel. If `NULL` (default), selection starts from the set of
+#'        unpenalized covariates.
 #' @param out.csv If not `NULL`, the name of a CSV file to save the
 #'        output to.
 #'
 #' @return
 #' A data frame of class `projsel` where each row corresponds to a
 #' forward-selected submodel that contains all variables listed up to that row.
+#' Attribute `start.from` reports the predictors in the initial model.
 #' The data frame contains the following columns:
 #' \item{var}{names of the variables selected.}
 #' \item{kl}{KL-divergence from the full model to the submodel.}
 #' \item{rel.kl.null}{relative explanatory power of predictors starting from the
 #'       intercept-only model.}
-#' \item{rel.kl}{relative explanatory power of predictors starting from the model
-#'       containing unpenalized covariates.}
+#' \item{rel.kl}{relative explanatory power of predictors starting from the
+#'       initial model.}
 #' \item{elpd}{the expected log predictive density of the submodels.}
 #' \item{delta.elpd}{the difference in elpd from the full model.}
 #'
@@ -184,7 +188,8 @@ fit.submodel <- function(x, sigma2, mu, chosen, xt, yt, logistic) {
 #'
 #' @importFrom utils write.csv
 #' @export
-projsel <- function(obj, max.iters=30, out.csv=NULL) {
+projsel <- function(obj, max.iters=30, start.from=NULL,
+                    out.csv=NULL) {
     validate.hsstan(obj)
     validate.samples(obj)
     if (length(obj$model.terms$penalized) == 0)
@@ -194,9 +199,16 @@ projsel <- function(obj, max.iters=30, out.csv=NULL) {
     yt <- obj$data[[obj$model.terms$outcome]]
     is.logistic <- is.logistic(obj)
     sigma2 <- if (is.logistic) 1 else as.matrix(obj$stanfit, pars="sigma")^2
-    U <- length(obj$betas$unpenalized)
-    K <- length(obj$betas$penalized)
-    P <- U + K
+
+    ## set of variables in the initial submodel
+    start.idx <- validate.start.from(obj, start.from)
+    chosen <- start.idx
+
+    ## number of model parameters (including intercept)
+    P <- sum(sapply(obj$betas, length))
+
+    ## number of iterations to run
+    K <- min(P - length(start.idx), max.iters)
 
     message(sprintf("%58s  %8s %11s", "Model", "KL", "ELPD"))
     report.iter <- function(msg, kl, elpd)
@@ -210,11 +222,12 @@ projsel <- function(obj, max.iters=30, out.csv=NULL) {
     kl.elpd <- c(sub$kl, sub$elpd)
     report.iter("Intercept only", sub$kl, sub$elpd)
 
-    ## start from the model having only unpenalized variables
-    chosen <- 1:U
-    sub <- fit.submodel(x, sigma2, fit, chosen, xt, yt, is.logistic)
-    kl.elpd <- rbind(kl.elpd, c(sub$kl, sub$elpd))
-    report.iter("Unpenalized covariates", sub$kl, sub$elpd)
+    ## initial submodel with the set of chosen covariates
+    if (length(start.idx) > 1) {
+        sub <- fit.submodel(x, sigma2, fit, chosen, xt, yt, is.logistic)
+        kl.elpd <- rbind(kl.elpd, c(sub$kl, sub$elpd))
+        report.iter("Initial submodel", sub$kl, sub$elpd)
+    }
 
     ## add variables one at a time
     for (iter in 1:K) {
@@ -225,9 +238,6 @@ projsel <- function(obj, max.iters=30, out.csv=NULL) {
         sub <- fit.submodel(x, sigma2, fit, chosen, xt, yt, is.logistic)
         kl.elpd <- rbind(kl.elpd, c(sub$kl, sub$elpd))
         report.iter(colnames(x)[sel.idx], sub$kl, sub$elpd)
-
-        if (iter == max.iters)
-            break
     }
 
     ## evaluate the full model if selection stopped before reaching it
@@ -236,14 +246,16 @@ projsel <- function(obj, max.iters=30, out.csv=NULL) {
 
     kl <- kl.elpd[, 1]
     res <- data.frame(var=c("Intercept only",
-                            "Unpenalized covariates",
-                            colnames(x)[setdiff(chosen, 1:U)]),
+                            if (length(start.idx) > 1) "Initial submodel",
+                            colnames(x)[setdiff(chosen, start.idx)]),
                       kl=kl,
                       rel.kl.null=1 - kl / kl[1],
                       rel.kl=c(NA, 1 - kl[-1] / kl[2]),
                       elpd=kl.elpd[, 2],
                       delta.elpd=kl.elpd[, 2] - full$elpd,
                       stringsAsFactors=FALSE)
+    attr(res, "start.from") <- if (!is.null(start.from))
+                                   start.from else obj$model.terms$unpenalized
     if (!is.null(out.csv))
         write.csv(file=out.csv, res, row.names=FALSE)