+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Add the `sub.idx` parameter to `posterior_performance()` to select the
   observations to be used in the computation of the performance measures.
+- Add the `start.from` option to run `projsel()` to start the selection
+  procedure from a smaller submodel than the set of unpenalized covariates.
 
 ### Smaller Changes and Bug Fixes